use Volume::operator() with long args.
git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/anatomist/anatomist-free/trunk@49928 613412ce-1361-4b7a-93bf-a3c75af1bae4

+++ src/volrender/object/volrender.cc
@@ -448,7 +448,7 @@ namespace
         || d->zscalefac != 1 )
     {
       vol = VolumeRef<T>( d->texdimx, d->texdimy, d->texdimz );
-      size_t x, y, z, t = t0;
+      long x, y, z, t = t0;
       /* cout << "resampling to " << d->dimx << "/" << d->texdimx << ", "
           << d->dimy << "/" << d->texdimy << ", "
           << d->dimz << "/" << d->texdimz << endl; */
@@ -510,7 +510,7 @@ namespace
         || d->zscalefac != 1 || scl != 1. || offset != 0. )
     {
       vol = VolumeRef<U>( d->texdimx, d->texdimy, d->texdimz );
-      size_t x, y, z, t = t0;
+      long x, y, z, t = t0;
       /* cout << "resampling to " << d->dimx << "/" << d->texdimx << ", "
       << d->dimy << "/" << d->texdimy << ", "
       << d->dimz << "/" << d->texdimz << endl; */
@@ -555,7 +555,7 @@ namespace
     rc_ptr<Volume<AimsRGB> > v0 = avol->volume()->volume();
     VolumeRef<AimsRGBA>  vol = VolumeRef<AimsRGBA>( d->texdimx, d->texdimy,
         d->texdimz );
-    size_t x, y, z, t = t0;
+    long x, y, z, t = t0;
     /* cout << "resampling to " << d->dimx << "/" << d->texdimx << ", "
         << d->dimy << "/" << d->texdimy << ", "
         << d->dimz << "/" << d->texdimz << endl; */
